requirements.txt
discord.py[voice]==2.7.1
davey==0.1.4
PyNaCl==1.6.2
yt-dlp